feat(standard_layout): Refactor feeds to internal/
