--- handler_feed.go
@@ -1,53 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"time"
-
-	"github.com/google/uuid"
-	"github.com/mike-kimani/fechronizo/internal/database"
-)
-
-func (apiCfg *apiConfig) handlerCreateFeed(w http.ResponseWriter, r *http.Request, user database.User) {
-
-	type parameters struct {
-		Name string `json:"name"`
-		URL  string `json:"url"`
-	}
-
-	decoder := json.NewDecoder(r.Body)
-
-	params := parameters{}
-	err := decoder.Decode(&params)
-	if err != nil {
-		respondWithError(w, 400, fmt.Sprintf("Error parsing json: %v", err))
-		return
-	}
-
-	feed, err := apiCfg.DB.CreateFeeds(r.Context(), database.CreateFeedsParams{
-		ID:        uuid.New(),
-		CreatedAt: time.Now().UTC(),
-		UpdatedAt: time.Now().UTC(),
-		Name:      params.Name,
-		Url:       params.URL,
-		UserID:    user.ID,
-		UserName:  user.Name,
-	})
-	if err != nil {
-		respondWithError(w, 400, fmt.Sprintf("Couldn't create feed: %v", err))
-		return
-	}
-
-	respondWithJson(w, 201, databaseFeedToFeed(feed))
-}
-
-func (apiCfg *apiConfig) handlerGetFeed(w http.ResponseWriter, r *http.Request) {
-
-	feeds, err := apiCfg.DB.GetFeeds(r.Context())
-	if err != nil {
-		respondWithError(w, 400, fmt.Sprintf("Couldn't get feeds %v", err))
-	}
-	respondWithJson(w, 200, databaseFeedstoFeeds(feeds))
-}
